resolving ::Rails conflict
